nirtfeatures: Added NI RT features driver
Added an NI RT features driver. This is an ACPI device that exposes LEDs,
switches, and other hardware features of the Smasher controllers.

Not all of the proposed features of the device work as expected, and some
features may be removed in the future. Development work on this device by
the hardware team is currently not a high priority. These issues will be
addressed once the hardware team gets back to this device.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>
[gratian: fix conflict with 7a6ff4c4cbc3 ("misc: hisi_hikey_usb: Driver to support onboard USB gpio hub on Hikey960")]
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
[gratian: fix conflict with bb3b6552a5b0 ("staging: hikey9xx: split hi6421v600 irq into a separate driver")]
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
[gratian: fix conflict with f396ededbd8b ("misc: open-dice: Add driver to expose DICE data to userspace")]
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
[gratian: fix trivial conflict with 6c93c6f3bad4 ("misc: Add a mechanism to detect stalls on guest vCPUs")]
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
[gratian: fix trivial conflict with 393fc2f5948f ("misc: microchip: pci1xxxx: load auxiliary bus driver for the PIO function in the multi-function endpoint of pci1xxxx device.")]
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>