8250: Do not create ttyS* nodes for nonexistant devices
This commit is a merge of two upstream-unfriendly commits:
   commit ba93f2d13b34 ("8250: Make SERIAL_8250_RUNTIME_UARTS work correctly")
   commit 0d1ed2d7adc1 ("Revert "serial: 8250: Do nothing if nr_uarts=0"")

This was an attempt to resolve an architectural oddity in the 8250 core
code, which was that, by design, it will create a static (build- or
cmdline-specified) number of ttyS* devices, which is at odds with how
most other Linux kernel drivers work. The original commit messages did
not accurately describe the problem, so this is a merge-and-replace in
order to get a better commit description. So, without further ado:

The current state of affairs is as follows:

There exists two build-time constants.
   - CONFIG_SERIAL_8250_NR_UARTS: the build-time max number of 8250 UARTs.
   - CONFIG_SERIAL_8250_RUNTIME_UARTS: the boot-time number of 8250 UARTs.
        (confusingly, this is also the default setting of 8250.nr_uarts)

The way that the 8250 driver handles creation of ttyS* device is that, at
initialize time, we create min(8250.nr_uarts, NR_UARTS) ttyS* devices
up-front, and when a UART is registered, serial8250_find_match_or_unused
will either merge it in under an existing ttyS* or find an unused one.

There also exists code in the 8250 driver to automatically enumerate the
UARTs specified in the SERIAL_PORT_DFNS table; these are the "well-known"
ports at 0x3F8/0x2F8/0x3E8/0x2E8 from the IBM PC platform, and they are
added this way because they predate the use of ACPI tables for enumeration.

(Oh, and also you can point the ttyS* devices to completely different
addresses from userspace via TIOCSSERIAL.)

This has two ramifications for NI devices:

1) There is a limit on number of UARTs with 8250_core; this is a noteworthy
   problem when you are also a vendor of multiport serial cards, such as
   the PXIe-8430/16. However, the user experience of increasing the limit
   with a generic platform kernel is also bad; if, as a distro, we release
   a kernel with a default limit of 128, then users that have PXI systems
   that don't have any multiport serial cards will still be presented with
   devices named ttyS0 through ttyS127, even though most of them are just
   "empty" slots that are not backed by real devices.

2) ttyS0 through ttyS3 will always be given the addresses for the "legacy"
   serial ports, even if on-board controller ports are at different
   addresses. This is the case on several NI controllers; for example, on
   some sbRIO products [1] the UARTs are at 0x3F8, 0x350, 0x360, etc.
   However, because the legacy ports get first-dibs, those ports will be
   ttyS0, ttyS4, ttyS5, ... because the legacy 0x2f8/0x3E8/0x2E8 get S1-S3.

Our initial attempt at resolving this was to redefine 8250.nr_uarts to mean
"the number of automatically-created UART entries" and define it to be 0
(because we neither want the legacy ports nor the extra "empty" ports). On
NI platforms, all serial ports are described in the ACPI table, so legacy
enumeration is neither necessary nor desirable. However, this approach got
soundly rejected upstream because it "caused existing kernel configurations
to act differently from before" [2].

We're not alone in finding the upstream behavior to be counterintuitive;
I have found at least two other attempts [3, 4] to resolve it with similar
pushback of the form "this breaks existing users".

A "proper" upstreamable solution probably needs two parts, with the default
settings such that the current behavior is retained but can be opted-out
(via Kconfig and/or kernel cmdline).
- a config token for "don't create a bunch of empty devices" (1)
- a config token for "don't add legacy ports, let ACPI handle it" (2)

Until that arrives, we're stuck keeping this around, because we're _also_
stuck with not wanting to renumber ttyS* devices from the way we've shipped
them.

[1] https://github.com/ni/meta-nilrt/blob/8106f31da6980ee4ee94fa0e03b991479d9aa43e/recipes-kernel/kernel-tests/kernel-tests-files/test_kernel_serial_devices.sh#L127
[2] https://lore.kernel.org/lkml/20130603213754.GA15479@kroah.com/
[3] https://lore.kernel.org/linux-serial/1420513785-23660-1-git-send-email-peter@hurleysoftware.com/
[4] https://lore.kernel.org/linux-serial/20221025073944.102437-1-martin@geanix.com/

[SBOs from initial patches]
Natinst-CAR-ID: 634278
Signed-off-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Natinst-ReviewBoard-ID: 183619

Upstream-Status: Denied [rejected upstream]
Natinst-AzDO-ID: 2133864
Signed-off-by: Brenda Streiff <brenda.streiff@ni.com>